♻️ Refactor: GradientBackground
