+++ client/src/features/artist/components/Artist.tsx
@@ -1,6 +1,6 @@
 import ArtistTable from './ArtistTable.tsx';
 import ArtistHeader from './ArtistHeader.tsx';
-import GradientBackground from '../../../components/GradientBackground/GradientBackground.tsx';
+import GradientBackground from '../../../ui/GradientBackground/GradientBackground.tsx';
 import { useNavigate, useParams } from 'react-router-dom';
 import { useAppDispatch, useAppSelector } from '../../../app/hooks.ts';
 import { useEffect } from 'react';
