+++ client/src/ui/GradientBackground/GradientBackground.tsx
@@ -11,19 +11,18 @@ const GradientBackground = ({
   color,
   children,
 }: GradientBackgroundProps) => {
-  return (
-    <div
-      className={className}
-      style={{
-        backgroundImage: `linear-gradient(rgba(23, 23, 23, 0.5), rgb(23, 23, 23) 40vh),
+  const style = {
+    backgroundImage: `linear-gradient(rgba(23, 23, 23, 0.5), rgb(23, 23, 23) 40vh),
         linear-gradient(
         ${color},
         ${color} 40vh,
         transparent 40vh,
         transparent 100%
     )`,
-      }}
-    >
+  };
+
+  return (
+    <div className={className} style={style}>
       {children}
     </div>
   );
